requirements.txt
click==8.1.3
cloudpickle==2.2.0
commonmark==0.9.1
ConfigSpace==0.6.1
contourpy==1.0.6
cycler==0.11.0
dask==2023.5.1
dask-jobqueue==0.8.1
distributed==2023.5.1
emcee==3.1.4
fonttools==4.38.0
fsspec==2023.5.0
gym==0.21.0
importlib-metadata==4.13.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
locket==1.0.0
MarkupSafe==2.1.2
matplotlib==3.6.2
more-itertools==9.1.0
mpmath==1.2.1
msgpack==1.0.5
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
pandas==1.5.2
partd==1.4.0
Pillow==9.3.0
psutil==5.9.5
pyaml==21.10.1
Pygments==2.13.0
pynisher==1.0.5
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrfr==0.9.0
pyserial==3.5
python-dateutil==2.8.2
python-vxi11==0.9
pytz==2022.6
PyYAML==6.0
regex==2023.5.5
rich==12.6.0
scikit-learn==1.2.0
scikit-optimize==0.9.0
scipy==1.9.3
six==1.16.0
slicer==0.0.7
smac==2.0.1
sortedcontainers==2.4.0
stable-baselines3==1.6.2
sympy==1.11.1
git+ssh://git@github.com/decryptofy/debug-renesas-fmpi-stm32f0.git@9edc93e1b20e17df8a5124b7268cac321e1ca7ad#egg=TargetDriver
tblib==1.7.0
threadpoolctl==3.1.0
toolz==0.12.0
torch==1.13.0
tornado==6.3.2
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==2.0.2
zict==3.0.0
zipp==3.11.0
